update with Baxter going back
Saturday Baxter decided he didn't want be cooped up in the house
anymore,
he decided to make a break for it, out the front door. Our best guess is
that
once he was out he went into "hunting mode." Unfortunately one of the
neighborhood Pomeranian's almost did not survive that experience. Due to
this
experience and many other experinces over the last year. We have decided
that Baxter is not a good fit for our home and neighborhood, we took him
back to the shelter earlier today. It was one of the hardest decisons we
have
made in our marriage. That being said we feel it is best for all
involved. I have
included a few pictures in memory of our, crazy, wonderful, diffictult,
and
cherished Baxter.